// backend/doctor-service/src/encounter/dto/encounterResponse.dto.ts
export class EncounterResponseDto {
  id: string;
  status: string;
  patient: {
    reference: string;
    display: string;
  };
  practitioner: {
    reference: string;
    display: string;
  };
  period: {
    start: string;
    end?: string;
  };
  location?: string;
  reasonCodes: string[];
  lastUpdated: string;
}